+++ packages/pixel-craft/src/system/animator/system.ts
@@ -13,6 +13,47 @@ export class AnimatorSystem implements System {
 
   constructor(private readonly timer: TimerSystem) {}
 
+  static createAnimated(
+    options: Pick<Animated, 'animations' | 'transitions'>,
+  ): Animated {
+    const entryTransition = options.transitions.find(
+      (transition) => transition.from.type === TransitionType.Entry,
+    );
+    if (!entryTransition) {
+      throw new Error('No entry transition found!');
+    }
+    const animation = options.animations[entryTransition.to];
+
+    return {
+      animationFrame: 0,
+      animationTimer: 0,
+      animation,
+      animations: options.animations,
+      transitions: options.transitions,
+      possibleTransitions: AnimatorSystem.findPossibleTransitions(
+        options.transitions,
+        animation,
+      ),
+    };
+  }
+
+  static findPossibleTransitions(
+    transitions: Animated['transitions'],
+    newAnimation: Animation,
+  ): Animated['transitions'] {
+    return transitions.filter((transition) => {
+      if (transition.from.type === TransitionType.Any) {
+        return (
+          transition.to !== newAnimation.name || newAnimation.interruptible
+        );
+      }
+      return (
+        transition.from.type === TransitionType.Animation &&
+        transition.from.animation === newAnimation.name
+      );
+    });
+  }
+
   addGameObject(sprite: unknown): void {
     if (sprite && typeof sprite === 'object' && 'animations' in sprite) {
       const animatedSprite = sprite as Sprite & Animated;
@@ -64,24 +105,13 @@ export class AnimatorSystem implements System {
   }
 
   animate(sprite: Sprite & Animated, animation: Animation): void {
-    this.findPossibleTransitions(sprite, animation);
+    sprite.possibleTransitions = AnimatorSystem.findPossibleTransitions(
+      sprite.transitions,
+      animation,
+    );
     sprite.animation = animation;
     sprite.animationFrame = 0;
     sprite.animationTimer = 0;
     sprite.frame = sprite.animation.animationFrames[sprite.animationFrame];
-  }
-
-  private findPossibleTransitions(sprite: Animated, newAnimation: Animation) {
-    sprite.possibleTransitions = sprite.transitions.filter((transition) => {
-      if (transition.from.type === TransitionType.Any) {
-        return (
-          transition.to !== newAnimation.name || newAnimation.interruptible
-        );
-      }
-      return (
-        transition.from.type === TransitionType.Animation &&
-        transition.from.animation === newAnimation.name
-      );
-    });
   }
 }